Refactor to better fit the design
